+++ e2e-tests/contentful/src/pages/json.js
@@ -7,6 +7,8 @@ import Layout from "../components/layout"
 const JSONPage = ({ data }) => {
   const simple = data.simple.json
   const actors = data.complex.json.Actors
+  const english = data.english.jsonLocalized
+  const german = data.german.jsonLocalized
   return (
     <Layout>
       <h1>Simple:</h1>
@@ -26,22 +28,30 @@ const JSONPage = ({ data }) => {
               margin: "1rem 0",
             }}
           >
-            <p data-cy-value-name>Name: {actor.name}</p>
-            <p data-cy-value-photo>Photo: {actor.photo}</p>
-            <p data-cy-value-birthdate>Birthdate: {actor.Birthdate}</p>
-            <p data-cy-value-born-at>Born at: {actor.Born_At}</p>
-            <p data-cy-value-weight>Weight: {actor.weight}</p>
-            <p data-cy-value-age>Age: {actor.age}</p>
-            <p data-cy-value-wife>Wife: {actor.wife}</p>
-            <p data-cy-value-children>Children: {actor.children.join(", ")}</p>
-            <p data-cy-value-has-children>
-              Has children: {JSON.stringify(actor.hasChildren)}
-            </p>
-            <p data-cy-value-has-grey-hair>
-              Has grey hair: {JSON.stringify(actor.hasGreyHair)}
-            </p>
+            <p>Name: {actor.name}</p>
+            <p>Photo: {actor.photo}</p>
+            <p>Birthdate: {actor.Birthdate}</p>
+            <p>Born at: {actor.Born_At}</p>
+            <p>Weight: {actor.weight}</p>
+            <p>Age: {actor.age}</p>
+            <p>Wife: {actor.wife}</p>
+            <p>Children: {actor.children.join(", ")}</p>
+            <p>Has children: {JSON.stringify(actor.hasChildren)}</p>
+            <p>Has grey hair: {JSON.stringify(actor.hasGreyHair)}</p>
           </div>
         ))}
+      </div>
+      <h1>English:</h1>
+      <div data-cy-id="english">
+        <p data-cy-value-name>Name: {english.name}</p>
+        <p data-cy-value-city>City: {english.city}</p>
+        <p data-cy-value-age>Age: {english.age}</p>
+      </div>
+      <h1>German:</h1>
+      <div data-cy-id="german">
+        <p data-cy-value-name>Name: {german.name}</p>
+        <p data-cy-value-city>City: {german.city}</p>
+        <p data-cy-value-age>Age: {german.age}</p>
       </div>
     </Layout>
   )
@@ -72,6 +82,28 @@ export const pageQuery = graphql`
           hasChildren
           hasGreyHair
         }
+      }
+    }
+    english: contentfulJson(
+      node_locale: { eq: "en-US" }
+      jsonLocalized: { id: { ne: null } }
+    ) {
+      title
+      jsonLocalized {
+        age
+        city
+        name
+      }
+    }
+    german: contentfulJson(
+      node_locale: { eq: "de-DE" }
+      jsonLocalized: { id: { ne: null } }
+    ) {
+      title
+      jsonLocalized {
+        age
+        city
+        name
       }
     }
   }