+++ packages/gatsby-plugin-google-tagmanager/src/__tests__/gatsby-browser.js
@@ -1,3 +1,23 @@
+import { getLCP, getFID, getCLS } from "web-vitals/base"
+
+jest.mock(`web-vitals/base`, () => {
+  function createEntry(type, id, value) {
+    return { name: type, id, value }
+  }
+
+  return {
+    getLCP: jest.fn(report => {
+      report(createEntry(`LCP`, `1`, `300`))
+    }),
+    getFID: jest.fn(report => {
+      report(createEntry(`FID`, `2`, `150`))
+    }),
+    getCLS: jest.fn(report => {
+      report(createEntry(`CLS`, `3`, `0.10`))
+    }),
+  }
+})
+
 const getAPI = setup => {
   if (setup) {
     setup()
@@ -8,10 +28,17 @@ const getAPI = setup => {
   return require(`../gatsby-browser`)
 }
 
+let currentNodeEnv
 beforeEach(() => {
+  currentNodeEnv = process.env.NODE_ENV
   window.dataLayer = []
   jest.useFakeTimers()
   process.env.NODE_ENV = undefined
+})
+
+afterEach(() => {
+  jest.useRealTimers()
+  process.env.NODE_ENV = currentNodeEnv
 })
 
 describe(`onRouteUpdate`, () => {
@@ -109,5 +136,61 @@ describe(`onRouteUpdate`, () => {
     jest.runAllTimers()
 
     expect(window[dataLayerName]).toHaveLength(1)
+  })
+
+  it(`sends core web vitals when enabled`, async () => {
+    const { onInitialClientRender } = getAPI(() => {
+      process.env.NODE_ENV = `production`
+    })
+    onInitialClientRender({}, { enableWebVitalsTracking: true })
+
+    // wait 2 ticks to wait for dynamic import to resolve
+    await Promise.resolve()
+    await Promise.resolve()
+
+    jest.runAllTimers()
+
+    expect(window.dataLayer.length).toBe(3)
+    expect(window.dataLayer).toContainEqual({
+      event: `core-web-vitals`,
+      webVitalsMeasurement: {
+        name: `LCP`,
+        id: `1`,
+        value: 300,
+      },
+    })
+    expect(window.dataLayer).toContainEqual({
+      event: `core-web-vitals`,
+      webVitalsMeasurement: {
+        name: `FID`,
+        id: `2`,
+        value: 150,
+      },
+    })
+    expect(window.dataLayer).toContainEqual({
+      event: `core-web-vitals`,
+      webVitalsMeasurement: {
+        name: `CLS`,
+        id: `3`,
+        value: 100,
+      },
+    })
+  })
+
+  it(`sends nothing when web vitals tracking is disabled`, async () => {
+    const { onInitialClientRender } = getAPI(() => {
+      process.env.NODE_ENV = `production`
+    })
+    onInitialClientRender({}, { enableWebVitalsTracking: false })
+
+    // wait 2 ticks to wait for dynamic import to resolve
+    await Promise.resolve()
+    await Promise.resolve()
+
+    jest.runAllTimers()
+
+    expect(getLCP).not.toBeCalled()
+    expect(getFID).not.toBeCalled()
+    expect(getCLS).not.toBeCalled()
   })
 })