+++ packages/gatsby-source-shopify/src/processors/index.ts
@@ -1,0 +1,2 @@
+export * from "./collections"
+export * from "./incremental-products"