+++ packages/gatsby/src/internal-plugins/internal-data-bridge/gatsby-node.js
@@ -5,7 +5,7 @@ const _ = require(`lodash`)
 
 const { emitter, store } = require(`../../redux`)
 const { actions } = require(`../../redux/actions`)
-const { getNode } = require(`../../redux/nodes`)
+const { getNode } = require(`../../datastore`)
 
 function transformPackageJson(json) {
   const transformDeps = deps =>
@@ -119,46 +119,8 @@ exports.sourceNodes = ({
   watchConfig(pathToGatsbyConfig, createGatsbyConfigNode)
 
   // Create nodes for functions
-  if (process.env.GATSBY_EXPERIMENTAL_FUNCTIONS) {
-    const { functions } = store.getState()
-    const createFunctionNode = config => {
-      createNode({
-        id: `gatsby-function-${config.absoluteCompiledFilePath}`,
-        ...config,
-        parent: null,
-        children: [],
-        internal: {
-          contentDigest: createContentDigest(config),
-          type: `SiteFunction`,
-        },
-      })
-    }
-    functions.forEach(config => {
-      createFunctionNode(config)
-    })
-
-    // Listen for updates to functions to update the nodes.
-    emitter.on(`SET_SITE_FUNCTIONS`, action => {
-      // Identify any now deleted functions and remove their nodes.
-      const existingNodes = getNodesByType(`SiteFunction`)
-      const newFunctionsSet = new Set()
-      action.payload.forEach(config =>
-        newFunctionsSet.add(
-          `gatsby-function-${config.absoluteCompiledFilePath}`
-        )
-      )
-      const toBeDeleted = existingNodes.filter(
-        node => !newFunctionsSet.has(node.id)
-      )
-      toBeDeleted.forEach(node => deleteNode(node))
-
-      action.payload.forEach(config => {
-        createFunctionNode(config)
-      })
-    })
-  } else {
-    // If not enabled, create a dummy node so we can ignore it in the dev 404 page
-    const config = { functionRoute: `FAKE`, absoluteCompiledFilePath: `FAKE` }
+  const { functions } = store.getState()
+  const createFunctionNode = config => {
     createNode({
       id: `gatsby-function-${config.absoluteCompiledFilePath}`,
       ...config,
@@ -170,6 +132,27 @@ exports.sourceNodes = ({
       },
     })
   }
+  functions.forEach(config => {
+    createFunctionNode(config)
+  })
+
+  // Listen for updates to functions to update the nodes.
+  emitter.on(`SET_SITE_FUNCTIONS`, action => {
+    // Identify any now deleted functions and remove their nodes.
+    const existingNodes = getNodesByType(`SiteFunction`)
+    const newFunctionsSet = new Set()
+    action.payload.forEach(config =>
+      newFunctionsSet.add(`gatsby-function-${config.absoluteCompiledFilePath}`)
+    )
+    const toBeDeleted = existingNodes.filter(
+      node => !newFunctionsSet.has(node.id)
+    )
+    toBeDeleted.forEach(node => deleteNode(node))
+
+    action.payload.forEach(config => {
+      createFunctionNode(config)
+    })
+  })
 }
 
 function watchConfig(pathToGatsbyConfig, createGatsbyConfigNode) {