+++ packages/gatsby/src/services/create-pages.ts
@@ -5,6 +5,7 @@ import { assertStore } from "../utils/assert-store"
 import { IGatsbyPage } from "../redux/types"
 import { actions } from "../redux/actions"
 import { deleteUntouchedPages, findChangedPages } from "../utils/changed-pages"
+import { getDataStore } from "../datastore"
 
 export async function createPages({
   parentSpan,
@@ -57,18 +58,19 @@ export async function createPages({
     activity.end()
   }
 
+  const dataStore = getDataStore()
   reporter.info(
-    `Total nodes: ${store.getState().nodes.size}, SitePage nodes: ${
-      store.getState().nodesByType?.get(`SitePage`)?.size
-    } (use --verbose for breakdown)`
+    `Total nodes: ${dataStore.countNodes()}, ` +
+      `SitePage nodes: ${
+        store.getState().pages.size
+      } (use --verbose for breakdown)`
   )
 
   if (process.env.gatsby_log_level === `verbose`) {
+    const types = dataStore.getTypes()
     reporter.verbose(
-      `Number of node types: ${
-        store.getState().nodesByType.size
-      }. Nodes per type: ${[...store.getState().nodesByType.entries()]
-        .map(([type, nodes]) => type + `: ` + nodes.size)
+      `Number of node types: ${types.length}. Nodes per type: ${types
+        .map(type => type + `: ` + dataStore.countNodes(type))
         .join(`, `)}`
     )
   }
@@ -98,6 +100,7 @@ export async function createPages({
       changedPages.length === 1 ? `` : `s`
     }`
   )
+
   tim.end()
 
   store.dispatch(actions.apiFinished({ apiName: `createPages` }))