+++ packages/gatsby/src/utils/worker/__tests__/test-helpers/index.ts
@@ -1,0 +1,2 @@
+export * from "./create-test-worker"
+export * from "./jest-helpers"