+++ packages/gatsby-admin/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.19.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.18.0-next.2...gatsby-admin@0.19.0-next.0) (2021-06-18)
+
+**Note:** Version bump only for package gatsby-admin
+
+# [0.18.0-next.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.18.0-next.1...gatsby-admin@0.18.0-next.2) (2021-06-17)
+
+**Note:** Version bump only for package gatsby-admin
+
+# [0.18.0-next.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.18.0-next.0...gatsby-admin@0.18.0-next.1) (2021-06-10)
+
+**Note:** Version bump only for package gatsby-admin
+
+# [0.18.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.17.0-next.2...gatsby-admin@0.18.0-next.0) (2021-06-03)
+
+**Note:** Version bump only for package gatsby-admin
+
+# [0.17.0-next.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.17.0-next.1...gatsby-admin@0.17.0-next.2) (2021-06-02)
+
+**Note:** Version bump only for package gatsby-admin
+
+# [0.17.0-next.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.17.0-next.0...gatsby-admin@0.17.0-next.1) (2021-05-31)
+
+**Note:** Version bump only for package gatsby-admin
+
 # [0.17.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-admin@0.16.0-next.6...gatsby-admin@0.17.0-next.0) (2021-05-21)
 
 **Note:** Version bump only for package gatsby-admin