+++ packages/gatsby-plugin-google-analytics/CHANGELOG.md
@@ -3,6 +3,32 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.9.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-google-analytics@3.8.0-next.2...gatsby-plugin-google-analytics@3.9.0-next.0) (2021-06-18)
+
+**Note:** Version bump only for package gatsby-plugin-google-analytics
+
+# [3.8.0-next.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-google-analytics@3.8.0-next.1...gatsby-plugin-google-analytics@3.8.0-next.2) (2021-06-17)
+
+### Features
+
+- **gatsby-plugin-google-analytics:** enable core webvitals tracking ([#31665](https://github.com/gatsbyjs/gatsby/issues/31665)) ([1ecd6e1](https://github.com/gatsbyjs/gatsby/commit/1ecd6e12eeedcabc54f3be00137a5d092978de58))
+
+# [3.8.0-next.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-google-analytics@3.8.0-next.0...gatsby-plugin-google-analytics@3.8.0-next.1) (2021-06-10)
+
+**Note:** Version bump only for package gatsby-plugin-google-analytics
+
+# [3.8.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-google-analytics@3.7.0-next.2...gatsby-plugin-google-analytics@3.8.0-next.0) (2021-06-03)
+
+**Note:** Version bump only for package gatsby-plugin-google-analytics
+
+# [3.7.0-next.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-google-analytics@3.7.0-next.1...gatsby-plugin-google-analytics@3.7.0-next.2) (2021-06-02)
+
+**Note:** Version bump only for package gatsby-plugin-google-analytics
+
+# [3.7.0-next.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-google-analytics@3.7.0-next.0...gatsby-plugin-google-analytics@3.7.0-next.1) (2021-05-31)
+
+**Note:** Version bump only for package gatsby-plugin-google-analytics
+
 # [3.7.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-google-analytics@3.6.0-next.0...gatsby-plugin-google-analytics@3.7.0-next.0) (2021-05-21)
 
 **Note:** Version bump only for package gatsby-plugin-google-analytics