+++ packages/gatsby-source-npm-package-search/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.9.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-npm-package-search@3.8.0-next.1...gatsby-source-npm-package-search@3.9.0-next.0) (2021-06-18)
+
+**Note:** Version bump only for package gatsby-source-npm-package-search
+
+# [3.8.0-next.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-npm-package-search@3.8.0-next.0...gatsby-source-npm-package-search@3.8.0-next.1) (2021-06-10)
+
+### Bug Fixes
+
+- **deps:** update minor and patch for gatsby-source-npm-package-search ([#31704](https://github.com/gatsbyjs/gatsby/issues/31704)) ([29b8d0a](https://github.com/gatsbyjs/gatsby/commit/29b8d0ab1116ebff68f23fa9a05b07baefafe476))
+
+# [3.8.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-npm-package-search@3.7.0-next.2...gatsby-source-npm-package-search@3.8.0-next.0) (2021-06-03)
+
+**Note:** Version bump only for package gatsby-source-npm-package-search
+
+# [3.7.0-next.2](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-npm-package-search@3.7.0-next.1...gatsby-source-npm-package-search@3.7.0-next.2) (2021-06-02)
+
+**Note:** Version bump only for package gatsby-source-npm-package-search
+
+# [3.7.0-next.1](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-npm-package-search@3.7.0-next.0...gatsby-source-npm-package-search@3.7.0-next.1) (2021-05-31)
+
+**Note:** Version bump only for package gatsby-source-npm-package-search
+
 # [3.7.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-npm-package-search@3.6.0-next.0...gatsby-source-npm-package-search@3.7.0-next.0) (2021-05-21)
 
 **Note:** Version bump only for package gatsby-source-npm-package-search